+++ semcore/chart/src/Tooltip/index.tsx
@@ -25,7 +25,7 @@ export interface IDefaultTooltip {
 }
 
 /**
- * @deprecated Please, use package `@semcore/d3-chart` instead. Package `@semcore/chart` will be removed in the next major release
+ * @deprecated Please, use package `@semcore/ui/d3-chart` instead. Package `@semcore/chart` is deprecated.
  */
 class DefaultTooltipContent extends PureComponent<IDefaultTooltip> {
   static displayName = 'DefaultTooltipContent';