+++ semcore/neighbor-location/src/NeighborLocation.jsx
@@ -1,100 +1,118 @@
-import React, { useContext, useRef } from 'react';
-import createComponent, { Root, Component, register, CONTEXT_COMPONENT } from '@semcore/core';
+import React, { useContext } from 'react';
+import createComponent, { Root, Component, register } from '@semcore/core';
 import getOriginChildren from '@semcore/utils/lib/getOriginChildren';
 import isNode from '@semcore/utils/lib/isNode';
 
-const CALCULATE_NEIGHBOR_LOCATION = Symbol('CALCULATE_NEIGHBOR_LOCATION');
+const Context = register.get('neighbor-location-context', React.createContext({}));
+
+function childrenWithoutFragment(children) {
+  return React.Children.toArray(children).reduce((acc, node) => {
+    if (React.isValidElement(node) && node.type === React.Fragment) {
+      acc.push.apply(acc, childrenWithoutFragment(node.props.children));
+    } else {
+      acc.push(node);
+    }
+    return acc;
+  }, []);
+}
+
+function calculateNeighborLocation(length, i) {
+  // default state
+  let neighborLocation = 'both';
+  // if one child
+  if (length === 1) {
+    return undefined;
+  }
+  // if first child
+  if (i === 0) {
+    neighborLocation = 'right';
+  }
+  // if last child
+  if (i === length - 1) {
+    neighborLocation = 'left';
+  }
+  return neighborLocation;
+}
 
 class NeighborLocationRoot extends Component {
   static displayName = 'NeighborLocation';
 
-  _currentIndex = null;
+  controlsLengthRef = React.createRef();
+  cacheChild = new Map();
 
-  calculateNeighborLocation(controlsLength, prevNeighborLocation) {
-    // default state
-    let neighborLocation = 'both';
-
-    if (this._currentIndex === null) {
-      // for re-render component
-      if (prevNeighborLocation) {
-        // return prev state
-        return prevNeighborLocation;
-      } else {
-        // for not context component
-        return undefined;
-      }
-    }
-    // if one children
-    if (controlsLength === 1) {
-      return undefined;
-    }
-    // if first children
-    if (this._currentIndex === 0) {
-      neighborLocation = 'right';
-    }
-    // if last children
-    if (this._currentIndex === controlsLength - 1) {
-      neighborLocation = 'left';
-    }
-    // increment index children
-    this._currentIndex = this._currentIndex + 1;
-    // if finished render all children
-    if (this._currentIndex === controlsLength) {
-      // reset index
-      this._currentIndex = null;
+  calculateNeighborLocation(controlsLength = 0, component) {
+    // for not re-render component
+    if (!this.cacheChild.has(component)) {
+      const neighborLocation = calculateNeighborLocation(controlsLength, this.cacheChild.size);
+      // set cache
+      this.cacheChild.set(component, neighborLocation);
     }
     // return default state
-    return neighborLocation;
+    return this.cacheChild.get(component);
+  }
+
+  getDetectProps() {
+    return {
+      getNeighborLocation: this.calculateNeighborLocation.bind(
+        this,
+        this.controlsLengthRef.current,
+      ),
+    };
   }
 
   setContext() {
-    const {
-      Children,
-      controlsLength = React.Children.toArray(getOriginChildren(Children)).filter(isNode).length,
-    } = this.asProps;
-
     return {
-      [CALCULATE_NEIGHBOR_LOCATION]: this.calculateNeighborLocation.bind(this, controlsLength),
+      controlsLengthRef: this.controlsLengthRef,
     };
   }
 
   render() {
-    const { Children, tag: Tag } = this.asProps;
-    this._currentIndex = 0;
+    const { Children, tag: Tag, controlsLength } = this.asProps;
+    this.controlsLengthRef.current =
+      controlsLength ?? childrenWithoutFragment(getOriginChildren(Children)).filter(isNode).length;
+    this.cacheChild.clear();
 
-    if (Tag) {
-      return <Root render={Tag}>{Children.origin}</Root>;
-    }
+    if (Tag)
+      return (
+        <Root render={Tag}>
+          <Children />
+        </Root>
+      );
 
-    return Children.origin;
+    return <Children />;
   }
 }
 
-function neighborLocationEnhance() {
-  return (props) => {
-    const prevNeighborLocationRef = useRef(undefined);
-    const ctx = useContext(NeighborLocation[CONTEXT_COMPONENT]);
-    if (ctx[CALCULATE_NEIGHBOR_LOCATION]) {
-      prevNeighborLocationRef['current'] = ctx[CALCULATE_NEIGHBOR_LOCATION](
-        prevNeighborLocationRef['current'],
-      );
-    }
-    return {
-      neighborLocation: prevNeighborLocationRef['current'],
-      ...props,
-    };
-  };
+class Detect extends Component {
+  render() {
+    const { children, neighborLocation: selfNeighborLocation, getNeighborLocation } = this.asProps;
+    const calculateNeighborLocation = getNeighborLocation ? getNeighborLocation(this) : undefined;
+    const neighborLocation = selfNeighborLocation ?? calculateNeighborLocation;
+
+    if (!children) return;
+    if (typeof children === 'function') return children(neighborLocation);
+    return React.cloneElement(React.Children.only(children), {
+      neighborLocation,
+    });
+  }
 }
 
-const Context = register.get('neighbor-location-context', React.createContext({}));
+function useNeighborLocationDetect(index) {
+  const { controlsLengthRef } = useContext(Context);
+  if (controlsLengthRef.current === undefined) return false;
+  return calculateNeighborLocation(controlsLengthRef.current, index);
+}
+
 const NeighborLocation = createComponent(
   NeighborLocationRoot,
-  {},
+  {
+    Detect: Detect,
+  },
   {
     context: Context,
   },
 );
 
-export { neighborLocationEnhance };
+export { useNeighborLocationDetect };
 
 export default NeighborLocation;