+++ semcore/pills/src/Pills.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import createComponent, { Component, sstyled, Root } from '@semcore/core';
 import { Box } from '@semcore/flex-box';
-import NeighborLocation, { neighborLocationEnhance } from '@semcore/neighbor-location';
+import NeighborLocation, { useNeighborLocationDetect } from '@semcore/neighbor-location';
 import keyboardFocusEnhance from '@semcore/utils/lib/enhances/keyboardFocusEnhance';
 import addonTextChildren from '@semcore/utils/lib/addonTextChildren';
 
@@ -25,9 +25,10 @@ class RootPills extends Component {
     this.handlers.value(value, e);
   };
 
-  getItemProps(props) {
+  getItemProps(props, i) {
     const { value, size, disabled } = this.asProps;
     return {
+      index: i,
       size,
       disabled,
       selected: value === props.value,
@@ -51,14 +52,15 @@ class RootPills extends Component {
 
 function Pill(props) {
   const SPill = Root;
-  const { Children, styles, addonLeft, addonRight, selected, disabled } = props;
-
+  const { Children, styles, addonLeft, addonRight, selected, disabled, index } = props;
+  const neighborLocation = useNeighborLocationDetect(index);
   return sstyled(styles)(
     <SPill
       render={Box}
-      type="button"
       tag="button"
+      type="button"
       role="radio"
+      neighborLocation={neighborLocation}
       aria-checked={selected}
       aria-disabled={disabled}
     >
@@ -69,7 +71,7 @@ function Pill(props) {
   );
 }
 
-Pill.enhance = [keyboardFocusEnhance(), neighborLocationEnhance()];
+Pill.enhance = [keyboardFocusEnhance()];
 
 function Text(props) {
   const SText = Root;