+++ tools/jest-preset-ui/jest-preset.js
@@ -16,13 +16,15 @@ module.exports = {
     '^@semcore/utils/lib/(.*)': '@semcore/utils/src/$1',
     '^@semcore/(.*)/lib/(.*)': '@semcore/$1/lib/$2',
     '^@semcore/icon/(.*)/(.*)': '@semcore/icon/$1/$2',
+    // TODO: this is temporary crunch
+    '^@semcore/icon$': resolvePath(`${__dirname}/../../semcore/icon/src`),
     '^@semcore/(.*)': '@semcore/$1/src',
     '^disable-jest-mapper:(.*)': '$1',
-    '^react$': resolvePath(__dirname, '../../node_modules/react/umd/react.production.min.js'),
-    '^react-dom$': resolvePath(
-      __dirname,
-      '../../node_modules/react-dom/umd/react-dom.production.min.js',
-    ),
+    // '^react$': resolvePath(__dirname, '../../node_modules/react/umd/react.production.min.js'),
+    // '^react-dom$': resolvePath(
+    //   __dirname,
+    //   '../../node_modules/react-dom/umd/react-dom.production.min.js',
+    // ),
   },
   modulePathIgnorePatterns: ['/generator-component/'],
   testEnvironment: 'jsdom',