+++ website/docs/patterns/feedback-yes-no/examples/feedback-yes-no.jsx
@@ -33,9 +33,8 @@ const validate = {
 };
 
 class Feedback extends React.PureComponent {
-  handleChange = (fn, trigger) => (value, e) => {
+  handleChange = (fn) => (value, e) => {
     fn(e);
-    this.props.onChange(e, trigger);
   };
 
   render() {
@@ -58,7 +57,7 @@ class Feedback extends React.PureComponent {
                   {...input}
                   autoFocus
                   h={80}
-                  onChange={this.handleChange(input.onChange, 'description')}
+                  onChange={this.handleChange(input.onChange)}
                   id="suggestions"
                 />
               )}
@@ -71,11 +70,7 @@ class Feedback extends React.PureComponent {
             <FeedbackForm.Item name="email" validate={validate.email}>
               {({ input }) => (
                 <Input state={input.state}>
-                  <Input.Value
-                    {...input}
-                    onChange={this.handleChange(input.onChange, 'email')}
-                    id="email"
-                  />
+                  <Input.Value {...input} onChange={this.handleChange(input.onChange)} id="email" />
                 </Input>
               )}
             </FeedbackForm.Item>