+++ website/docs/table-group/data-table/examples/sticky.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useRef } from 'react';
 import DataTable from '@semcore/data-table';
-import Sticky from '@semcore/sticky';
+import { Box } from '@semcore/flex-box';
 import ScrollArea from '@semcore/scroll-area';
 
 export default () => {
@@ -16,19 +16,19 @@ export default () => {
   return (
     <>
       <DataTable data={data}>
-        <Sticky zIndex={2} top={top}>
+        <Box position="sticky" zIndex={2} top={top}>
           <DataTable.Head wMin={1000}>
             <DataTable.Column name="keyword" children="Keyword" />
             <DataTable.Column name="kd" children="KD,%" />
             <DataTable.Column name="cpc" children="CPC" />
             <DataTable.Column name="vol" children="Vol." />
           </DataTable.Head>
-        </Sticky>
+        </Box>
         <DataTable.Body />
       </DataTable>
       <h3>with Scroll.Bar in Header</h3>
       <DataTable data={data}>
-        <Sticky zIndex={2} top={top}>
+        <Box position="sticky" zIndex={2} top={top}>
           <DataTable.Head wMin={1000} ref={containerRef}>
             <DataTable.Column name="keyword" children="Keyword" />
             <DataTable.Column name="kd" children="KD,%" />
@@ -36,7 +36,7 @@ export default () => {
             <DataTable.Column name="vol" children="Vol." />
           </DataTable.Head>
           {container && <ScrollArea.Bar container={container} />}
-        </Sticky>
+        </Box>
         <DataTable.Body />
       </DataTable>
     </>