+++ semcore/data-table/CHANGELOG.md
@@ -2,7 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [3.1.15] - 2022-10-06
+## [3.3.0] - 2022-10-10
+
+### Changed
+
+- Added support for React 18 🔥
+- Extended version range for dependency `@semcore/icons`.
+
+## [3.2.0] - 2022-10-07
+
+### Added
+
+- Added support `ref` for `DataTable.Column` and `DataTable.Cell`.
 
 ### Changed